% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_agency_crime}
\alias{get_agency_crime}
\title{Title}
\usage{
get_agency_crime(agency_ori = NULL, since = NULL, until = NULL)
}
\arguments{
\item{agency_ori}{(character vector), denoting the agency ori, could be looked up by get_agencies()
default: NULL, which means getting all agencies}

\item{since}{(integer), starting year of query}

\item{until}{(integer), ending year of query}
}
\value{
tibble::tibble
}
\description{
Title
}
\details{
since >= 1985, until <= 2018

(https://www.ucrdatatool.gov/)

national crime estimates from 1960 through the most recent year available

state crime estimates from 1960 through the most recent year available

city and county crime counts from 1985 through the most recent year available
}
\examples{
\dontrun{
as <- get_agency_crime("AK0010100", since = 1984, until = 2018)
asny <- get_agency_crime("NY330SS00", since = 1985, until = 2018)
}
}
