% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_agency_participation}
\alias{get_agency_participation}
\title{Title}
\usage{
get_agency_participation(ori)
}
\arguments{
\item{ori}{(character) agency ori, an identifier defined by API}
}
\value{
a tibble::tibble of pariticipation info of agencies
}
\description{
Title
}
\examples{
\dontrun{
agency_par <- get_agency_participation("AK0010100")
}
}
