% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_url}
\alias{get_url}
\title{The most basic function in crimer}
\usage{
get_url(endpoint)
}
\arguments{
\item{endpoint}{(character) endpoint defined by (https://crime-data-explorer.fr.cloud.gov/api)}
}
\value{
NULL if encounter error in requesting,
otherwise tibble::tibble from result JSON
}
\description{
The most basic function in crimer
}
\examples{
# this is equivalent to get_agencies() function
# in fact, get_agencies() is defined by doing so
agencies <- get_url("api/agencies/list")
}
